feat: Add icon for app in flavor module
This commit adds an icon for the app in the flavor module. The icon file "ic_launcher.png" is included in the assets/images directory. This addition enhances the visual representation of the app in different flavors.

Note: This commit message follows the conventional format of starting with a type (feat) and a brief description of the changes.