chore: Refactor runCommand function to use cli_spinner package
This commit refactors the `runCommand` function in the `command.dart` file to use the `cli_spinner` package for displaying loading spinners. The `cli_spinner` package provides a simple and customizable way to show loading spinners in the command-line interface. By using this package, we can improve the user experience by visually indicating that a process is running. This change enhances the readability and maintainability of the code.

Note: This commit message follows the conventional format of starting with a type (chore) and a brief description of the changes.